Modify display behavior
As it stood previously, the display function only progressively reduced
the font size of the expression with the font size only resetting when
the clear button was pressed. Replace this with a mutationObserver that
sizes the font relative to the number of digits that will fit in the
available space.