Fix two bugs
The first bug is an issue with the expression retaining a decimal if an
operator was pressed before any number came after the decimal. Fix
validator to replace the decimal in this case. The second bug is other
buttons being activated if the mouse clicks down on one button and is
released on another. Change the condition in mousedown() to resolve
this.